refactor: 상품 리스트 페이지 포맷 수정
